Rewards Page - track goal/tasks milestones

+++ app/src/main/java/com/example/goal_garden_project/models/GoalWithPlantPicture.kt
@@ -5,8 +5,8 @@ import androidx.room.Relation
 
 class GoalWithPlantPicture(
     val title: String,
-    val pictureId: Int,
-    val plantId: Int,
+    val pictureId: Long,
+    val plantId: Long,
     val progressionStage: Int,
     val imageUrl: String,
     val goalId: Int
